refactor(posts): reorganize package structure for category and location data sources
