+++ features/posts/common/data/src/commonMain/kotlin/com/parksupark/soomjae/features/posts/common/data/category/repositories/CategoryRepositoryImpl.kt
@@ -1,12 +1,12 @@
-package com.parksupark.soomjae.features.posts.common.data.repositories
+package com.parksupark.soomjae.features.posts.common.data.category.repositories
 
 import arrow.core.Either
 import com.parksupark.soomjae.core.domain.failures.DataFailure
-import com.parksupark.soomjae.features.posts.common.data.sources.CategoryRemoteSource
+import com.parksupark.soomjae.features.posts.common.data.category.sources.CategoryRemoteSource
 import com.parksupark.soomjae.features.posts.common.domain.models.Category
 import com.parksupark.soomjae.features.posts.common.domain.repositories.CategoryRepository
 
-class CategoryRepositoryImpl(
+internal class CategoryRepositoryImpl(
     private val categoryRemoteSource: CategoryRemoteSource,
 ) : CategoryRepository {
     override suspend fun getAllCategories(): Either<DataFailure.Network, Map<Long, Category>> = categoryRemoteSource.getAllCategories()
