+++ features/posts/common/data/src/commonMain/kotlin/com/parksupark/soomjae/features/posts/common/data/category/sources/CategoryRemoteSource.kt
@@ -1,4 +1,4 @@
-package com.parksupark.soomjae.features.posts.common.data.sources
+package com.parksupark.soomjae.features.posts.common.data.category.sources
 
 import arrow.core.Either
 import com.parksupark.soomjae.core.domain.failures.DataFailure
