+++ features/posts/common/data/src/commonMain/kotlin/com/parksupark/soomjae/features/posts/common/data/di/FeaturesPostsCommonDataModule.kt
@@ -1,26 +1,26 @@
 package com.parksupark.soomjae.features.posts.common.data.di
 
-import com.parksupark.soomjae.features.posts.common.data.repositories.CategoryRepositoryImpl
-import com.parksupark.soomjae.features.posts.common.data.repositories.DefaultLocationRepository
-import com.parksupark.soomjae.features.posts.common.data.sources.CategoryRemoteSource
-import com.parksupark.soomjae.features.posts.common.data.sources.CategoryRemoteSourceImpl
-import com.parksupark.soomjae.features.posts.common.data.sources.DefaultRemoteLocationSourceImpl
-import com.parksupark.soomjae.features.posts.common.data.sources.RemoteCommunityPostLikeSource
-import com.parksupark.soomjae.features.posts.common.data.sources.RemoteCommunityPostLikeSourceImpl
-import com.parksupark.soomjae.features.posts.common.data.sources.RemoteLocationSource
+import com.parksupark.soomjae.features.posts.common.data.category.repositories.CategoryRepositoryImpl
+import com.parksupark.soomjae.features.posts.common.data.category.sources.CategoryRemoteSource
+import com.parksupark.soomjae.features.posts.common.data.category.sources.CategoryRemoteSourceImpl
+import com.parksupark.soomjae.features.posts.common.data.like.sources.RemoteCommunityPostLikeSource
+import com.parksupark.soomjae.features.posts.common.data.like.sources.RemoteCommunityPostLikeSourceImpl
+import com.parksupark.soomjae.features.posts.common.data.location.repositories.DefaultLocationRepository
+import com.parksupark.soomjae.features.posts.common.data.location.sources.DefaultRemoteLocationSourceImpl
+import com.parksupark.soomjae.features.posts.common.data.location.sources.RemoteLocationSource
 import com.parksupark.soomjae.features.posts.common.domain.repositories.CategoryRepository
 import com.parksupark.soomjae.features.posts.common.domain.repositories.LocationRepository
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.bind
 import org.koin.dsl.module
 
-private val sourcesModule = module {
+private val categoryModule = module {
     singleOf(::CategoryRemoteSourceImpl).bind<CategoryRemoteSource>()
-    singleOf(::RemoteCommunityPostLikeSourceImpl).bind<RemoteCommunityPostLikeSource>()
+    singleOf(::CategoryRepositoryImpl).bind<CategoryRepository>()
 }
 
-private val repositoriesModule = module {
-    singleOf(::CategoryRepositoryImpl).bind<CategoryRepository>()
+private val likeModule = module {
+    singleOf(::RemoteCommunityPostLikeSourceImpl).bind<RemoteCommunityPostLikeSource>()
 }
 
 private val locationModule = module {
@@ -29,5 +29,5 @@ private val locationModule = module {
 }
 
 val featuresPostsCommonDataModule = module {
-    includes(sourcesModule, repositoriesModule, locationModule)
+    includes(categoryModule, likeModule, locationModule)
 }
