+++ features/posts/common/data/src/commonMain/kotlin/com/parksupark/soomjae/features/posts/common/data/location/repositories/DefaultLocationRepository.kt
@@ -1,9 +1,9 @@
-package com.parksupark.soomjae.features.posts.common.data.repositories
+package com.parksupark.soomjae.features.posts.common.data.location.repositories
 
 import arrow.core.Either
 import com.parksupark.soomjae.core.domain.failures.DataFailure
-import com.parksupark.soomjae.features.posts.common.data.dtos.toLocation
-import com.parksupark.soomjae.features.posts.common.data.sources.RemoteLocationSource
+import com.parksupark.soomjae.features.posts.common.data.location.dtos.toLocation
+import com.parksupark.soomjae.features.posts.common.data.location.sources.RemoteLocationSource
 import com.parksupark.soomjae.features.posts.common.domain.models.Location
 import com.parksupark.soomjae.features.posts.common.domain.repositories.LocationRepository
