+++ features/posts/common/data/src/commonMain/kotlin/com/parksupark/soomjae/features/posts/common/data/location/sources/RemoteLocationSource.kt
@@ -1,8 +1,8 @@
-package com.parksupark.soomjae.features.posts.common.data.sources
+package com.parksupark.soomjae.features.posts.common.data.location.sources
 
 import arrow.core.Either
 import com.parksupark.soomjae.core.domain.failures.DataFailure
-import com.parksupark.soomjae.features.posts.common.data.dtos.LocationResponse
+import com.parksupark.soomjae.features.posts.common.data.location.dtos.LocationResponse
 
 internal interface RemoteLocationSource {
     suspend fun getAllLocations(): Either<DataFailure.Network, List<LocationResponse>>
