+++ features/posts/community/data/src/commonMain/kotlin/com/parksupark/soomjae/features/posts/community/data/repositories/OfflineFirstCommunityPostLikeRepository.kt
@@ -2,7 +2,7 @@ package com.parksupark.soomjae.features.posts.community.data.repositories
 
 import arrow.core.Either
 import com.parksupark.soomjae.core.domain.failures.DataFailure
-import com.parksupark.soomjae.features.posts.common.data.sources.RemoteCommunityPostLikeSource
+import com.parksupark.soomjae.features.posts.common.data.like.sources.RemoteCommunityPostLikeSource
 import com.parksupark.soomjae.features.posts.common.domain.models.Like
 import com.parksupark.soomjae.features.posts.community.domain.repositories.CommunityPostLikeRepository
 import kotlinx.coroutines.flow.Flow
