+++ features/posts/meeting/data/src/commonMain/kotlin/com/parksupark/soomjae/features/posts/common/data/dtos/MeetingPostResponse.kt
@@ -1,5 +1,9 @@
 package com.parksupark.soomjae.features.posts.common.data.dtos
 
+import com.parksupark.soomjae.features.posts.common.data.category.dtos.CategoryResponse
+import com.parksupark.soomjae.features.posts.common.data.category.dtos.toDomain
+import com.parksupark.soomjae.features.posts.common.data.common.dtos.MemberResponse
+import com.parksupark.soomjae.features.posts.common.data.common.dtos.toModel
 import com.parksupark.soomjae.features.posts.common.domain.models.MeetingPost
 import kotlin.time.ExperimentalTime
 import kotlinx.datetime.LocalDateTime
